Changing health from int to float

+++ Assets/Scripts/PlayerHealth.cs
@@ -6,8 +6,8 @@ public class PlayerHealth : MonoBehaviour{
 	private static GameManager gameManager;
 	public static GameObject GameController;
 
-	public int startingHealth;                                  // The amount of health the player starts the game with.
-	public int currentHealth;                                   // The current health the player has.
+	public float startingHealth;                                  // The amount of health the player starts the game with.
+	public float currentHealth;                                   // The current health the player has.
 	public UnityEngine.UI.Slider healthSlider;                  // Reference to the UI's health bar.
 	public Image damageImage;                                   // Reference to an image to flash on the screen on being hurt.
 	public AudioClip deathClip;                                 // The audio clip to play when the player dies.
